asn1: pkcs15_object: generalize handling of EXPLICIT tags
subClassAttributes and typeAttributes are EXPLICITly tagged due
to X.680 30.6c. In pkcs15 objects these always contain a SEQUENCE
so it simplifies code to handle this in common code. Should there
ever be a need for anything else it can be indicated with
an additional field in struct sc_asn1_pkcs15_object.

This also fixes asn1 decoding bug for GenericSecretKey and
decoding/enconding bug for DataContainerObjects for which the
EXPLICIT SEQUENCE tag was not included for [1] TypeAttributes
portion of the PKCS15Object.

We may need to write some backwards compatibility parsing code
for the above encondings if they are deemed widely used.